pass userDelegate to function
